fix(maintainability): consolidate profile command ownership
